Create get_tstatXY.R
A pre-process script to get common and aligned SNPs between the summary statistics of two traits. 
The columns are renamed to match those used in LHC-MR, missing columns such as sample size are added, duplicate SNPs are removed by only keeping the ones with a closer MAF to the UK10K MAF in the LD file, the strands are then aligned, and the twi new trait files are saved.